refactor: split LegacySettings.tsx into settingsFields, PathMappingEditor, InstanceEditor
- settingsFields.ts (~241 LOC): FieldConfig interface, FIELDS array, NAV_GROUPS, TAB_KEYS, TABS
- PathMappingEditor.tsx (~207 LOC): path mapping CRUD + test-path widget
- InstanceEditor.tsx (~130 LOC): multi-instance editor for Sonarr/Radarr
- LegacySettings.tsx reduced to 682 LOC (from 1248) — imports from new modules
- FieldConfig and NAV_GROUPS re-exported for backwards compatibility
